Update attention.py 9th and 11th line.
The results will be same though you've made this change.
In the constructor of the Self Attention class,
in the 9th line of code, you're calculating self.head_dim by embed_size * heads. Instead of that, we can do that calculation by self.embed_size*self.head.
This holds True for the 11th Line of code also.

This change makes your code easier to read & understand.